Desenvolvida funcionalidade de importação de arquivo

+++ src/BancoBr.Common/Attributes/CampoCNABAttribute.cs
@@ -8,32 +8,32 @@ namespace BancoBr.Common.Attributes
     public class CampoCNABAttribute : Attribute
     {
         public bool Desabilitado { get; }
-        public double Posicao { get; }
+        public int Posicao { get; }
         public int Tamanho { get; }
         public char? CharPreenchimento { get; }
         public string Formatacao { get; }
 
-        public CampoCNABAttribute(double posicao, int tamanho)
+        public CampoCNABAttribute(int posicao, int tamanho)
         {
             Posicao = posicao;
             Tamanho = tamanho;
         }
 
-        public CampoCNABAttribute(double posicao, int tamanho, string formatacao)
+        public CampoCNABAttribute(int posicao, int tamanho, string formatacao)
         {
             Posicao = posicao;
             Tamanho = tamanho;
             Formatacao = formatacao;
         }
 
-        public CampoCNABAttribute(double posicao, int tamanho, char charPreenchimento)
+        public CampoCNABAttribute(int posicao, int tamanho, char charPreenchimento)
         {
             Tamanho = tamanho;
             Posicao = posicao;
             CharPreenchimento = charPreenchimento;
         }
 
-        public CampoCNABAttribute(double posicao, int tamanho, char charPreenchimento, string formatacao)
+        public CampoCNABAttribute(int posicao, int tamanho, char charPreenchimento, string formatacao)
         {
             Tamanho = tamanho;
             Posicao = posicao;
